+++ src/extension.ts
@@ -2,6 +2,7 @@
 // Import the module and reference it with the alias vscode in your code below
 import * as vscode from "vscode";
 import provideDefinition from "./provideDefinition";
+import provideMacroCompletionItems from "./provideMacroCompletionItems";
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
@@ -12,14 +13,30 @@ export function activate(context: vscode.ExtensionContext) {
     'Congratulations, your extension "vscode-freemarker-tipster" is now active!'
   );
 
+  const ftlDocumentSelector: vscode.DocumentSelector = {
+    language: "ftl",
+    scheme: "file",
+  };
+
+  // definition provider for `macro` and `function` directives
   const definitionProvider = vscode.languages.registerDefinitionProvider(
-    { language: "ftl", scheme: "file" },
+    ftlDocumentSelector,
     {
       provideDefinition,
     }
   );
 
-  context.subscriptions.push(definitionProvider);
+  // completion item provider for `macro` directives
+  const macroCompletionItemProvider =
+    vscode.languages.registerCompletionItemProvider(
+      ftlDocumentSelector,
+      {
+        provideCompletionItems: provideMacroCompletionItems,
+      },
+      "@"
+    );
+
+  context.subscriptions.push(definitionProvider, macroCompletionItemProvider);
 }
 
 // This method is called when your extension is deactivated